Refactor `Parser` class to simplify the API
