+++ src/library/classes/CellParser.ts
@@ -1,6 +1,5 @@
 // Library
-import { Parser } from ".";
-import { _Parser, ParserConstructorOptions } from "./_base";
+import { _Parser } from "./Parser";
 
 // -----
 // VSCSV
@@ -15,15 +14,7 @@ import { _Parser, ParserConstructorOptions } from "./_base";
  * 
  * @see {@link Cell}
  */
-export class VSCSV extends Parser<Cell> {
-
-    /**
-     * Instantiates a new {@link VSCSV}
-     * @param delimiter The delimiter to use (default: `,`)
-     */
-    constructor(opts?: Partial<ParserConstructorOptions>) {
-        super(opts);
-    }
+export class VSCSV extends _Parser<Cell> {
 
     protected parseCell(value: string, columnNumber: number, lineNumber: number): Cell {
         return {
@@ -33,24 +24,6 @@ export class VSCSV extends Parser<Cell> {
             columnEnd: columnNumber + value.length,
             toString: () => value
         };
-    }
-
-    protected parseLine(line: string, lineNumber: number): Cell[] {
-        return line
-            .split(this.delimiter)
-            .map((value, columnNumber) => {
-                return this.parseCell(value, columnNumber, lineNumber);
-            });
-    }
-
-    protected serializeCell(cell: Cell): string {
-        return cell.value;
-    }
-
-    protected serializeLine(cells: Cell[]): string {
-        return cells
-            .map(this.serializeCell)
-            .join(this.delimiter);
     }
 
 }
@@ -71,5 +44,6 @@ type Cell = {
     column: number;
     /** The column number of the cell's end */
     columnEnd: number;
+    /** Converts the cell into a string */
     toString: () => string;
 };
